Ensure initial tmp Nokogiri install links against system libs
We leverage Rubygems as an easy way to assemble a gem along with all of 
it's dependencies so a manifest can be generated for packaging time. 
Currently Rubygems does not provide a way to install a gem and NOT 
compile it's native extensions. We need this initial gem install to 
succeed even though we end up re-compiling all native extensions using 
the Junos SDK.

This commit ensures `libxml2` and `libxslt` are available on the build 
machine and forces the temporary Nokogiri install to link against these 
system level installs.